chore: (governance-aggr) CONTINUE

+++ aggregates/governance-aggr/src/main/java/com/bloxbean/cardano/yaci/store/governanceaggr/domain/LatestVotingProcedure.java
@@ -6,7 +6,6 @@ import com.bloxbean.cardano.yaci.store.common.domain.BlockAwareDomain;
 import com.fasterxml.jackson.annotation.JsonIgnoreProperties;
 import com.fasterxml.jackson.databind.PropertyNamingStrategies;
 import com.fasterxml.jackson.databind.annotation.JsonNaming;
-import jakarta.persistence.Column;
 import lombok.AllArgsConstructor;
 import lombok.Data;
 import lombok.NoArgsConstructor;
@@ -25,7 +24,7 @@ public class LatestVotingProcedure extends BlockAwareDomain {
 
     private String txHash;
 
-    private long index;
+    private Integer index;
 
     private Long slot;
